+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 2.0.3
 
 * [#93] Made it easier to set the error and connection listener on a stream connection
+* [#98] Added try-catch around subscription message callback
 
 ## Version 2.0.2